backend/requirements.txt
# Web framework
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
aiosqlite==0.19.0

# Data validation
pydantic==1.10.12

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# File handling & Security
python-magic==0.4.27
Pillow==10.1.0

# HTTP client
aiohttp==3.9.1

# Caching & Performance
redis==5.0.1
aiocache==0.12.2
asyncio-throttle==1.0.2
cachetools==5.3.2

# Monitoring & Profiling
psutil==5.9.6
memory-profiler==0.61.0
line-profiler==4.1.1
prometheus-client==0.19.0

# Utilities
tabulate==0.9.0 